@@ -1,13 +1,21 @@
+[build-system]
+requires = ["poetry_core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
+
 [tool.poetry]
 name = "pymeter"
 version = "0.1.0"
 description = ""
-authors = ["Kelvin.Ye <testmankelvin@163.com>"]
-readme = 'README.md'
-
+readme = "README.md"
+authors = [
+    "Kelvin.Ye <testmankelvin@163.com>"
+]
 packages = [
     { include = "pymeter" },
     { include = "pymeter/**/*.py" }
+]
+classifiers = [
+    "Programming Language :: Python"
 ]
 
 [[tool.poetry.source]]
@@ -21,46 +29,41 @@ url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
 priority = "primary"
 
 [tool.poetry.dependencies]
-python = "^3.8"
-jsonpath = "0.82.2"
+python = "^3.11"
 orjson = "3.9.7"
+jsonpath = "0.82.2"
 xmltodict = "0.13.0"
-sqlalchemy = "2.0.20"
+sqlalchemy = "2.0.21"
 cx-Oracle = "8.3.0"
-mysql-connector-python = "8.1.0"
-psycopg2-binary = "2.9.7"
 pyodbc = "4.0.39"
+psycopg2-binary = "2.9.7"
+mysql-connector-python = "8.1.0"
 sshtunnel = "0.4.0"
-pycryptodome = "3.18.0"
+pycryptodome = "3.19.0"
 python-engineio = "4.7.1"
 python-socketio = "5.9.0"
-httpx = "0.25.0"
-pytypes = "1.0b10"
 ulid-py = "1.1.0"
-bidict = "0.22.1"
-tabulate = "0.9.0"
-gevent = "23.9.1"
-uvloop = "0.17.0"
-loguru = "0.7.2"
 arrow = "1.2.3"
 faker = "19.6.1"
+httpx = "0.25.0"
 json5 = "0.9.14"
+bidict = "0.22.1"
+loguru = "0.7.2"
+gevent = "23.9.1"
+uvloop = "0.17.0"
+pytypes = "1.0b10"
+tabulate = "0.9.0"
 
 [tool.poetry.group.dev.dependencies]
 flask = "*"
 pytest = "*"
 pytest-html = "*"
-types-orjson = "*"
 ruff = "*"
 yapf = "*"
 black = "*"
 
-[build-system]
-requires = ["poetry_core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
-
 [tool.ruff]
-target-version = "py38"
+target-version = "py311"
 line-length = 120
 src = ["pymeter"]
 select = [
@@ -110,3 +113,20 @@ dedent_closing_brackets = true
 join_multiple_lines = true
 spaces_before_comment = 2
 split_before_dot = true
+
+[tool.pytest.ini_options]
+minversion = "6.0"
+testpaths = "tests"
+addopts = [
+    "--color=auto",
+    "--capture=tee-sys",
+    "--code-highlight=yes",
+    "--disable-warnings",
+    "--html=report.html",
+    "--self-contained-html"
+]
+log_cli = true
+log_level = "DEBUG"
+log_format = "[%(asctime)s][%(levelname)s][%(name)s.%(funcName)s:%(lineno)d] %(message)s"
+filterwarnings = ["error", "ignore::DeprecationWarning"]
+required_plugins = ["pytest-html"]
