@@ -2,6 +2,7 @@
 # @File    : gevent_test.py
 # @Time    : 2020/2/23 14:12
 # @Author  : Kelvin.Ye
+from datetime import UTC
 from datetime import datetime
 
 import gevent
@@ -15,13 +16,13 @@ coroutine_local = local()
 class Test(Greenlet):
     def _run(self):
         print('I am test')
-        # start_time = datetime.utcnow().strftime('%Y-%m-%d %H:%M:%S.%f')
-        # print(f'startTime={start_time}')
+        start_time = datetime.now(UTC).strftime('%Y-%m-%d %H:%M:%S.%f')
+        print(f'startTime={start_time}')
         coroutine_local.aa = 'aa'
         print(f'coroutine_local.aa={coroutine_local.aa}')
-        # gevent.sleep(1)
-        # end_time = datetime.utcnow().strftime('%Y-%m-%d %H:%M:%S.%f')
-        # print(f'endTime={end_time}')
+        gevent.sleep(1)
+        end_time = datetime.now(UTC).strftime('%Y-%m-%d %H:%M:%S.%f')
+        print(f'endTime={end_time}')
 
 
 if __name__ == '__main__':
