@@ -10,6 +10,6 @@ from pymeter.runner import Runner
 
 class TestRunner:
     def test_start(self):
-        with open(os.path.join(CONFIG.PROJECT_PATH, 'docs', 'test-script.json'), 'r', encoding='utf-8') as f:
+        with open(os.path.join(CONFIG.PROJECT_PATH, 'docs', 'test-script.json'), encoding='utf-8') as f:
             script = ''.join(f.readlines())
             Runner.start(script)
